Fix bugs from latest round of development.
